+++ examples/atomic_thread_pool_benchmark.rb
@@ -4,14 +4,6 @@ require "benchmark"
 require "concurrent-ruby"
 require_relative "../lib/atomic-ruby"
 
-def shareable_proc(&block)
-  if AtomicRuby::RACTOR_SAFE
-    Ractor.shareable_proc(&block)
-  else
-    block
-  end
-end
-
 results = []
 
 2.times do |idx|
@@ -22,11 +14,11 @@ results = []
     end
 
     100.times do
-      pool << shareable_proc { sleep(0.2) }
+      pool << proc { sleep(0.2) }
     end
 
     100.times do
-      pool << shareable_proc { 1_000_000.times.map(&:itself).sum }
+      pool << proc { 1_000_000.times.map(&:itself).sum }
     end
 
     pool.shutdown